test: refactor common unit tests
Refactor a series of common markup and state checks from the
deprecated 'builder' unit test into 'test_common' unit test (managed via
the 'test_main' suite).
