+++ test/unit-tests/common/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+__import__('pkg_resources').declare_namespace(__name__)
