+++ test/unit-tests/common/expected/admonitions.conf
@@ -1,5 +1,3 @@
-h1. Admonition Test
-
 {note}attention message{note}
 
 {warning}caution message{warning}
