+++ test/unit-tests/common/expected/formatting.conf
@@ -1,5 +1,3 @@
-h1. this is a text test
-
 _emphasis_
 
 *strong emphasis*
