+++ test/unit-tests/common/expected/list.conf
@@ -1,5 +1,3 @@
-h1. list test
-
 * BULLET 1
 * BULLET 2
