+++ test/unit-tests/common/expected/references.conf
@@ -1,5 +1,3 @@
-h1. references test
-
 [http://example.com/]
 
 [http://www.example.com/]
@@ -16,9 +14,9 @@ Leading text [http://example.com/] with trailing text.
 
 Leading text [Custom Link Name|http://example.com/] with trailing text.
 
-[TOCTREE]
+[base]
 
-[Custom Link Name|TOCTREE]
+[Custom Link Name|base]
 
 h1. sub&#45;section
