+++ test/unit-tests/common/expected/tables.conf
@@ -1,5 +1,3 @@
-h1. Table Test
-
 ||A||B||A or B||
 |False|False|False|
 |True|False|True|
